Add some camera stuff

+++ src/gamemodel/Constants.h
@@ -4,7 +4,7 @@
 namespace constants {
 	static const float TERRAIN_STEP_LENGTH = 1.0f;
 	static const float TERRAIN_HEIGHT_VARIATION_AMPLITUDE = 0.2f;
-	static const int   SCREEN_WIDTH = 15;
+	static const float SCREEN_WIDTH = 15.0f;
 	static const float MAN_HEIGHT = 2.0f;
 	static const float WORLD_INIT_LENGTH = 100.0f;
 };
